+++ build.sh
@@ -19,7 +19,7 @@ Options:
   -n, --no-build                  Skip the build step
   -v, --verbose                   Use normal verbosity (default: minimal)
   --use-prebuilt-assemblies       Build against pre-built Lidarr binaries
-  --lidarr-version VERSION        Override Lidarr assembly version (default: 3.1.3.4970)
+  --lidarr-version VERSION        Override Lidarr assembly version (default: 3.1.2.4913)
   -h, --help                      Show this help text
 
 Examples:
@@ -43,7 +43,7 @@ RESTORE=false
 NO_BUILD=false
 VERBOSE=false
 USE_PREBUILT=false
-LIDARR_VERSION="3.1.3.4970"
+LIDARR_VERSION="3.1.2.4913"
 
 while [[ $# -gt 0 ]]; do
   case "$1" in